+++ src/ops/redeem_logic_plugin_ops.rs
@@ -1,4 +1,4 @@
 pub mod redeem_logic_forward_ops;
 pub mod redeem_logic_settle_forward_ops;
 pub mod redeem_logic_vanilla_option_ops;
-pub mod redeem_logic_digital;
+pub mod redeem_logic_digital_ops;
