+++ src/ops/redeem_logic_plugin_ops/redeem_logic_digital_ops.rs
